interim commit: wrote Receiver but got side-tracked
The issue is that wrapping is getting in the way all over the place, it
is fragile and ugly. The only reason for it to exist is to transport a
failure response out of the handling of a Failed message, and that can
more reliably be done in a side-effecting fashion (like for
setReceiveTimeout). So, I’ll just stash this work and rip out Wrapper.