make FlexiRoute demand tracking type-safe
There is one caveat: in truth our variance annotations on
Source/Flow/Sink which need to be threaded all the way through to InPort
and OutPort are in fact not really correct (actually: inverted) when
writing the “inside” of a stream element, as an InPort behaves as a
source of things and should therefore be covariant. The only point where
this surfaces is the State (in both FlexiMerge/Route) which ties
together the port and element types. In the case of FlexiMerge things
work out by chance because type inference for varargs works covariantly,
“fixing” the wrong contravariance of InPort. For FlexiRoute there is no
such luck, scalac will happily accept all kinds of nonsense code due to
incorrect covariance. The problematic case is DemandFromAny which
demands GLB instead of LUB to be calculated. I fixed this by introducing
a mini-DSL that does this. It works like this:

    val select = p.out0 | p.out1 // will create a set with GLB of both
                                 // types and error out if none exists
    ...
    ctx.emit(select(port))(element)

where `port` is an untyped OutP that is found in the selector set using
object equality. The `port` argument is what the State gives the user,
forcing them to do this kind of resolution (they can also use `if` if
they want). Combine this with DemandFromAll not giving the user any port
info at all (since all of them have demand and can thus nicely be used
directly) and we got all the bases covered—hopefully.